+++ src/lib/sen/index.ts
@@ -10,3 +10,4 @@ export * from "./constants";
 export * from "./aggregate";
 export * from "./stats";
 export * from "./format";
+export * from "./calendar";